fixed chrome and windows 10 apps
Added additional chrome classname to float parameter. Fixed invisible
windows 10 apps being tiled by adding ApplicationFrameWindow to float
parameter. Note that this means windows 10 apps will not be tiled. This
could be addressed in the future by removing the classname from the
float parameter, and adding a check for invisible windows, or windows
with the classename but no title.